pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "job-scout"
version = "0.1.0"
description = "Daily job aggregation and email digest for PM roles"
requires-python = ">=3.11"
dependencies = [
    "python-jobspy>=1.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "feedparser>=6.0",
    "requests>=2.31",
]

[project.scripts]
job-scout = "job_scout.main:main"
